MGPMState

MGPMState

Construct an MGPM state vector.

MGPMState(s, ctx)

MGPMStateVector(K, n, l, r, m, v)

Details

The MGPMState of an MCMC is represented by a numerical vector of the form: $$\vec{s}=(K, n_2,...,n_K, l_2,...,l_{K+1}, r_2,...,r_{K+1}, m_1,...,m_R, v_1,...,v_P)^T$$ The element of this vector are described as follows:

K:

number of shifts;

\((n_2,...,n_{K+1}): \)

shift nodes - the shifts in the model occur at points within the branches leading to the shift nodes in tip-ward direction. The corresponding locations of these points are specified by \((l_2,...,l_{K+1})\). The nodes \((n_2,...,n_{K+1})\) should be ordered according to PCMTreePreorder(ctx$tree).

\((l_2,...,l_{K+1}): \)

offset of the shift points measured as distances from the beginnings of the branches leading to shift nodes in tip-ward direction.

\((r_2,...,r_{K+1}): \)

regime index vector. This is an integer vector with elements among \((N+1,n_2,...,n_{K+1})\), indicating the regime associated with each part in the tree. The regimes are named as the shift nodes, with N+1 corresponding to the part (and regime) originating at the root. It is possible to have lumped regimes, that is, different parts having the same regime. This regime-lumping must obey the following rules:

neighbor parts cannot have a lumped regime. Two parts originating at nodes \(n_i\) and \(n_j\) in the tree are called neighbor parts if they are separated solely by \(n_i\) or by \(n_j\);

to resolve the conflict between the shift nodes of the different parts covered by a lumped regime, it is established that the name of a lumped regime must equal the shift-node that appears first, according to PCMTreePreorder(ctx$tree).

\((m_1,...,m_R): \)

model type assignment to the unique regimes. This is an integer vector with elements between 1 and M, M denoting the total number of model types possible. Each element corresponds to an element in unique(c(N+1,r_2,...,r_{K+1}))

\((v_1,...,v_P):\)

real numbers passed to PCMParamLoadOrStore. This is a vectorized form of all model parameters.

Functions

MGPMStateVector: